requirements.txt
asgiref==3.7.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.2
Django==4.2.9
django-crispy-forms==1.7.2
idna==2.8
python-decouple==3.8
python-dotenv==1.0.0
pytz==2019.1
requests==2.31.0
six==1.12.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.16
